requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.2
anthropic>=0.40.0
